requirements.txt
pyspark==3.5.1
pandas 
numpy
scikit-learn
matplotlib
seaborn
streamlit
jupyter
plotly
altair
jupyterlab
requests
fastapi
uvicorn
websockets
psycopg2-binary
pymongo
apache-airflow==2.9.1
flask-session<0.6.0
